add resource operation requirement

+++ RestaurantAPI/Services/RestaurantService.cs
@@ -1,16 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-using AutoMapper;
-using Microsoft.AspNetCore.Mvc;
+﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
-using RestaurantAPI.Models;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
+using RestaurantAPI.Models;
+using System.Collections.Generic;
+using System.Linq;
+using System.Security.Claims;
 
 namespace RestaurantAPI.Services
 {
@@ -18,9 +16,9 @@ namespace RestaurantAPI.Services
     {
         RestaurantDto GetById(int id);
         IEnumerable<RestaurantDto> GetAll();
-        int Add(AddRestaurantDto dto);
-        void DeleteById(int id);
-        void UpdateById(int id, UpdateRestaurantDto restaurantDto);
+        int Add(AddRestaurantDto dto, int userId);
+        void DeleteById(int id, ClaimsPrincipal user);
+        void UpdateById(int id, UpdateRestaurantDto restaurantDto, ClaimsPrincipal user);
     }
 
     public class RestaurantService : IRestaurantService
@@ -28,12 +26,15 @@ namespace RestaurantAPI.Services
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
+        private readonly IAuthorizationService _authorizationService;
 
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger,
+            IAuthorizationService authorizationService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
         }
         public RestaurantDto GetById(int id)
         {
@@ -63,16 +64,19 @@ namespace RestaurantAPI.Services
             return result;
         }
 
-        public int Add(AddRestaurantDto dto)
+        public int Add(AddRestaurantDto dto, int userId)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+
+            restaurant.CreatedById = userId;
+
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
 
             return restaurant.Id;
         }
 
-        public void DeleteById(int id)
+        public void DeleteById(int id, ClaimsPrincipal user)
         {
             _logger.LogError($"Restaurant with id {id} DELETE action invoked");
             var restaurant = _dbContext.Restaurants.FirstOrDefault(r => r.Id == id);
@@ -80,15 +84,32 @@ namespace RestaurantAPI.Services
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
 
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, restaurant,
+                new ResourceOperationRequirement(ResourceOpetration.Delete)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException("Forbidden resource");
+            }
+
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
         }
 
-        public void UpdateById(int id, UpdateRestaurantDto restaurantDto)
+        public void UpdateById(int id, UpdateRestaurantDto restaurantDto, ClaimsPrincipal user)
         {
             var restaurant = _dbContext.Restaurants.FirstOrDefault(r => r.Id == id);
+
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, restaurant,
+                new ResourceOperationRequirement(ResourceOpetration.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException("Forbidden resource");
+            }
 
             restaurant.Name = restaurantDto.Name;
             restaurant.Description = restaurantDto.Description;
